gmail-integration/requirements.txt
# Gmail Integration Requirements
google-api-python-client==2.158.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
python-dotenv==1.0.1
requests==2.32.3
